Item ordering fix
Fixed an issue where items would be displayed out of order. This error
was caused by trying to honor the count feature in the JSON file. By
just listing duplicate items multiple times the bug has been fixed and
the game is fine with the duplicates. The count field also didn't allow
the order of the build to be correct unless the duplicate items were
supposed to be built one after another.