Starter Stacking
So in some cases stacking items makes sense. The best being for starter
items. Rather than show 3 health pots just stack them. For actual item
builds stacking does not make sense so it will still show duplicate
items. A character could in theory build 2 of the same items but not in
a row.