Progress Indicator, Readme, Experimental Save
Created a indicator to show x/x of current set download.  Updated the
readme to document the experimental save feature. If you specify the
absolute path to the /Config/Champions/ of your league of legends
install directory in riot_lol_config_champions.conf then the app will
auto save the files into the correct location for use with the league
client. THIS IS EXPERIMENTAL and is disabled by default.