Auto Save, Error Checking, Exe
Added the config button that allows the Riot Games path to be specified
so Item Sets are automatically saved in the correct location to be used
with the league client. Also added some error checking/reporting to http
get, file read and file write. Finally I am including the exe file to
make the setup process go away. May be removed depending on how github
reacts.