Settings Rewrite, UI Changes, Garena Supp Updated
There is now a main settings file for all the possible settings. This
required significant changes to the codebase. Also remove the help menu
and everything is now on the main screen. You can now specify the save
location for garena and league and then choose which one to use per
launch.